+++ jsonpath/function_extensions/match.py
@@ -1,50 +1,19 @@
 """The standard `match` function extension."""
 
-try:
-    import regex as re
-
-    REGEX_AVAILABLE = True
-except ImportError:
-    import re  # type: ignore
-
-    REGEX_AVAILABLE = False
-
-try:
-    from iregexp_check import check
-
-    IREGEXP_AVAILABLE = True
-except ImportError:
-    IREGEXP_AVAILABLE = False
-
-from jsonpath.function_extensions import ExpressionType
-from jsonpath.function_extensions import FilterFunction
-
-from ._pattern import map_re
+from ._pattern import AbstractRegexFilterFunction
 
 
-class Match(FilterFunction):
-    """A type-aware implementation of the standard `match` function."""
+class Match(AbstractRegexFilterFunction):
+    """The standard `match` function."""
 
-    arg_types = [ExpressionType.VALUE, ExpressionType.VALUE]
-    return_type = ExpressionType.LOGICAL
-
-    def __call__(self, string: str, pattern: str) -> bool:
-        """Return `True` if _string_ matches _pattern_, or `False` otherwise."""
-        # TODO: re.match caches compiled patterns internally, but `map_re` and `check`
-        # are not cached.
-
-        # TODO: validate literal patterns ar compile time?
-
-        if IREGEXP_AVAILABLE and (not isinstance(pattern, str) or not check(pattern)):
+    def __call__(self, value: object, pattern: object) -> bool:
+        """Return `True` if _value_ matches _pattern_, or `False` otherwise."""
+        if not isinstance(value, str) or not isinstance(pattern, str):
             return False
 
-        if REGEX_AVAILABLE:
-            try:
-                pattern = map_re(pattern)
-            except TypeError:
-                return False
+        _pattern = self.check_cache(pattern)
 
-        try:
-            return bool(re.fullmatch(pattern, string))
-        except (TypeError, re.error):
+        if _pattern is None:
             return False
+
+        return bool(_pattern.fullmatch(value))